feat(observability): オブザーバビリティ関連コードのリファクタリングと機能改善

+++ backend-api/internal/product/application/dto/get_image_response.go
@@ -2,13 +2,13 @@ package dto
 
 // GetImageResponse は画像取得のレスポンスDTO
 type GetImageResponse struct {
-	ProductID   int64  `json:"productId"`
+	ProductID   int    `json:"productId"`
 	ImageData   []byte `json:"-"`
 	ContentType string `json:"-"`
 }
 
 // NewGetImageResponse は新しいGetImageResponseを作成する
-func NewGetImageResponse(productID int64, imageData []byte, contentType string) *GetImageResponse {
+func NewGetImageResponse(productID int, imageData []byte, contentType string) *GetImageResponse {
 	return &GetImageResponse{
 		ProductID:   productID,
 		ImageData:   imageData,
