+++ backend-api/internal/product/application/dto/upload_image_request.go
@@ -2,13 +2,13 @@ package dto
 
 // UploadImageRequest は商品画像アップロードのリクエストDTO
 type UploadImageRequest struct {
-	ProductID int64
+	ProductID int
 	ImageData []byte
 	Filename  string
 }
 
 // NewUploadImageRequest は新しいUploadImageRequestを作成する
-func NewUploadImageRequest(productID int64, imageData []byte, filename string) *UploadImageRequest {
+func NewUploadImageRequest(productID int, imageData []byte, filename string) *UploadImageRequest {
 	return &UploadImageRequest{
 		ProductID: productID,
 		ImageData: imageData,
